+++ bench/run_linkbench/run.sh
@@ -5,10 +5,10 @@ maxid=$4
 dname=$5
 dop=$6
 secs=$7
-myORmo=$8
+dbms=$8
 dbhost=$9
 
-if [[ $myORmo = "mysql" ]]; then
+if [[ $dbms = "mysql" ]]; then
   echo Skip mstat
   # ps aux | grep python | grep mstat\.py | awk '{ print $2 }' | xargs kill -9 2> /dev/null
   # python mstat.py --loops 1000000 --interval 15 --db_user=root --db_password=pw --db_host=${dbhost} >& r.mstat.$fn &
@@ -32,39 +32,54 @@ du -sm --apparent-size $ddir >> r.asz1.$fn
 echo "before apparent $ddir" > r.asz2.$fn
 du -sm --apparent-size $ddir/* >> r.asz2.$fn
 
-if [[ $myORmo = "mongo" ]]; then
+if [[ $dbms = "mongo" ]]; then
   while :; do ps aux | grep mongod | grep -v grep; sleep 5; done >& r.ps.$fn &
   spid=$!
   props=LinkConfigMongoDb2.properties
   logarg=""
-else
+elif [[ $dbms = "mysql" ]]; then
   while :; do ps aux | grep mysqld | grep -v grep; sleep 5; done >& r.ps.$fn &
   spid=$!
   props=LinkConfigMysql.properties
   $client -uroot -ppw -A -h${dbhost} -e 'reset master'
   logarg="-Duser=root -Dpassword=pw"
+elif [[ $dbms = "postgres" ]]; then
+  while :; do ps aux | grep postgres | grep -v grep; sleep 5; done >& r.ps.$fn &
+  spid=$!
+  props=LinkConfigPgsql.properties
+  logarg="-Duser=linkbench -Dpassword=pw"
+else
+  echo dbms :: $dbms :: not supported
+  exit 1
 fi
 
 echo "background jobs: $ipid $vpid $spid" > r.o.$fn
 echo " config/${props} -Drequests=5000000000 -Drequesters=$dop -Dmaxtime=$secs -Dmaxid1=$maxid -Dprogressfreq=10 -Ddisplayfreq=10 -Dreq_progress_interval=100000 -Dhost=${dbhost} $logarg -Ddbid=linkdb0 -r" >> r.o.$fn
 
-/usr/bin/time -o r.time.$fn bash bin/linkbench -c config/${props} -Drequests=5000000000 -Drequesters=$dop -Dmaxtime=$secs -Dmaxid1=$maxid -Dprogressfreq=10 -Ddisplayfreq=10 -Dreq_progress_interval=100000 -Dhost=${dbhost} $logarg -Ddbid=linkdb0 -r >> r.o.$fn 2>&1
+if ! /usr/bin/time -o r.time.$fn bash bin/linkbench -c config/${props} -Drequests=5000000000 -Drequesters=$dop -Dmaxtime=$secs -Dmaxid1=$maxid -Dprogressfreq=10 -Ddisplayfreq=10 -Dreq_progress_interval=100000 -Dhost=${dbhost} $logarg -Ddbid=linkdb0 -r >> r.o.$fn 2>&1 ; then
+  echo Run failed
+  exit 1
+fi
 
 kill $ipid
 kill $vpid
 kill $spid
 # kill $mpid
 
-if [[ $myORmo = "mongo" ]]; then
+if [[ $dbms = "mongo" ]]; then
   cred="-u root -p pw --authenticationDatabase=admin"
-  echo "db.serverStatus()" | $client $cred linkdb0 > r.stat.$fn
-  echo "db.linktable.stats()" | $client $cred linkdb0 > r.link.$fn
-  echo "db.nodetable.stats()" | $client $cred linkdb0 > r.node.$fn
-  echo "db.counttable.stats()" | $client $cred linkdb0 > r.count.$fn
+  echo "db.serverStatus()" | $client $cred > r.srvstat.$fn
+  echo "db.stats()" | $client $cred > r.dbstats.$fn
+  echo "db.linktable.stats({indexDetails: true})" | $client $cred linkdb0 > r.stats.link.$fn
+  echo "db.nodetable.stats({indexDetails: true})" | $client $cred linkdb0 > r.stats.node.$fn
+  echo "db.counttable.stats({indexDetails: true})" | $client $cred linkdb0 > r.stats.count.$fn
+  echo "db.linktable.latencyStats({histograms: true}).pretty()" | $client $cred linkdb0 > r.lat.link.$fn
+  echo "db.nodetable.latencyStats({histograms: true}).pretty()" | $client $cred linkdb0 > r.lat.node.$fn
+  echo "db.counttable.latencyStats({histograms: true}).pretty()" | $client $cred linkdb0 > r.lat.count.$fn
   echo "db.oplog.rs.stats()" | $client $cred local > r.oplog.$fn
   echo "show dbs" | $client $cred linkdb0 > r.dbs.$fn
 
-else
+elif [[ $dbms = "mysql" ]]; then
   $client -uroot -ppw -A -h${dbhost} -e 'reset master'
   $client -uroot -ppw -A -h${dbhost} -e 'show engine innodb status\G' > r.esi.$fn
   $client -uroot -ppw -A -h${dbhost} -e 'show engine rocksdb status\G' > r.esr.$fn
@@ -76,6 +91,25 @@ else
   $client -uroot -ppw -A -h${dbhost} linkdb0 -e 'show indexes from nodetable' >> r.is.$fn
   $client -uroot -ppw -A -h${dbhost} linkdb0 -e 'show indexes from counttable' >> r.is.$fn
   $client -uroot -ppw -A -h${dbhost} -e 'show memory status\G' > r.mem.$fn
+  $client -uroot -ppw -A -h${dbhost} performance_schema -E -e 'SELECT * FROM table_io_waits_summary_by_table WHERE OBJECT_NAME IN ("linktable", "nodetable", "counttable")' > r.tstat1.$fn
+  $client -uroot -ppw -A -h${dbhost} performance_schema -E -e 'SELECT * FROM events_statements_summary_by_account_by_event_name WHERE USER="root"' > r.ustat1.$fn
+  $client -uroot -ppw -A -h${dbhost} -E -e 'SELECT * FROM sys.schema_table_statistics WHERE table_name IN ("linktable", "nodetable", "counttable")' > r.tstat2.$fn
+  $client -uroot -ppw -A -h${dbhost} -E -e 'SELECT * FROM sys.user_summary' > r.ustat2.$fn
+  $client -uroot -ppw -A -h${dbhost} -E -e 'SELECT * FROM information_schema.user_statistics' > r.ustat3.$fn
+  $client -uroot -ppw -A -h${dbhost} -E -e 'SELECT * FROM information_schema.table_statistics WHERE TABLE_NAME IN ("linktable", "nodetable", "counttable")' > r.tstat3.$fn
+  $client -uroot -ppw -A -h${dbhost} -E -e 'SELECT * FROM information_schema.index_statistics' > r.istat3.$fn
+elif [[ $dbms = "postgres" ]]; then
+  $client linkbench $pgauth -c 'show all' > o.pg.conf
+  $client linkbench $pgauth -x -c 'select * from pg_stat_bgwriter' > r.pgs.bg
+  $client linkbench $pgauth -x -c 'select * from pg_stat_database' > r.pgs.db
+  $client linkbench $pgauth -x -c 'select * from pg_stat_all_tables' > r.pgs.tabs
+  $client linkbench $pgauth -x -c 'select * from pg_stat_all_indexes' > r.pgs.idxs
+  $client linkbench $pgauth -x -c 'select * from pg_statio_all_tables' > r.pgi.tabs
+  $client linkbench $pgauth -x -c 'select * from pg_statio_all_indexes' > r.pgi.idxs
+  $client linkbench $pgauth -x -c 'select * from pg_statio_all_sequences' > r.pgi.seq
+else
+  echo dbms :: $dbms :: not supported
+  exit 1
 fi
 
 echo "after $ddir" >> r.sz1.$fn
